refactor: reusable single and multicard layouts
