+++ src/lib/components/layout/index.ts
@@ -4,6 +4,7 @@ export { default as Grid } from './grid.svelte';
 export { default as Subgrid } from './subgrid.svelte';
 export { default as Stack } from './stack.svelte';
 export { default as Switcher } from './switcher.svelte';
-export { default as PageColumns } from './pagecolumns.svelte';
+export { default as MultiCard } from './page/multicard.svelte';
+export { default as SingleCard } from './page/singlecard.svelte';
 export * as Sidebar from './sidebar';
 export * from './box';
